[YS-479] refactor: 참여자 자동 필터링 로직 분리
